+++ liquepy/__about__.py
@@ -1,4 +1,4 @@
-__version__ = "0.6.22"
+__version__ = "0.6.23"
 __project__ = "liquepy"
 __author__ = "Maxim Millen and Julieth Quintero"
 __license__ = "MIT"